# server/docs/swagger.yaml
basePath: /api
definitions:
  github_com_ai-comic-generator_server_internal_common.BaseResponse:
    properties:
      code:
        description: 业务状态码，0 表示成功，非 0 表示失败
        example: 0
        type: integer
      data:
        description: 响应数据，成功时返回；失败时通常为 null
      message:
        description: 响应消息，成功时为 ok，失败时为错误描述
        example: ok
        type: string
    type: object
  github_com_ai-comic-generator_server_internal_model.AddPayPackagePlanRequest:
    properties:
      amountFen:
        description: 售价（分）
        type: integer
      code:
        description: 套餐编码
        maxLength: 32
        minLength: 2
        type: string
      enabled:
        description: 上架状态（可选，默认 1）
        type: integer
      name:
        description: 展示名称
        maxLength: 64
        minLength: 1
        type: string
      points:
        description: 到账积分
        type: integer
      sortOrder:
        description: 排序（可选）
        type: integer
    required:
    - amountFen
    - code
    - name
    - points
    type: object
...
  github_com_ai-comic-generator_server_internal_model.AdminPayOrderPageRequest:
    properties:
      orderNo:
        description: 订单号模糊筛选
        type: string
      pageNum:
        description: 页码
        type: integer
      pageSize:
        description: 每页条数
        maximum: 100
        type: integer
      status:
        description: 订单状态筛选
        type: string
      userAccount:
        description: 登录账号模糊筛选
        type: string
      userId:
        description: 用户 ID 筛选
        type: integer
    required:
    - pageNum
    - pageSize
    type: object
...
  github_com_ai-comic-generator_server_internal_model.CreatePayOrderRequest:
    properties:
      channel:
        description: 支付渠道：alipay | mock
        type: string
      packageCode:
        description: 充值方案编码
        type: string
    required:
    - channel
    - packageCode
    type: object
...
  github_com_ai-comic-generator_server_internal_model.DeletePayPackagePlanRequest:
    properties:
      id:
        description: 方案主键 ID
        type: integer
    required:
    - id
    type: object
...
  github_com_ai-comic-generator_server_internal_model.PageResult:
    properties:
      pageNum:
        description: 当前页码
        example: 1
        type: integer
      pageSize:
        description: 每页条数
        example: 10
        type: integer
      records:
        description: 当前页数据列表（用户列表或文章列表）
      total:
        description: 符合条件的总条数
        example: 100
        type: integer
    type: object
...
  github_com_ai-comic-generator_server_internal_model.PayOrderPageRequest:
    properties:
      pageNum:
        description: 页码
        type: integer
      pageSize:
        description: 每页条数
        maximum: 100
        type: integer
    required:
    - pageNum
    - pageSize
    type: object
  github_com_ai-comic-generator_server_internal_model.PayPackagePlanPageRequest:
    properties:
      pageNum:
        description: 页码
        type: integer
      pageSize:
        description: 每页条数
        maximum: 100
        type: integer
    required:
    - pageNum
    - pageSize
    type: object
...
  github_com_ai-comic-generator_server_internal_model.UpdatePayPackagePlanRequest:
    properties:
      amountFen:
        description: 售价（分）
        type: integer
      enabled:
        description: 上架状态（可选）
        type: integer
      id:
        description: 方案主键 ID
        type: integer
      name:
        description: 展示名称
        maxLength: 64
        minLength: 1
        type: string
      points:
        description: 到账积分
        type: integer
      sortOrder:
        description: 排序（可选）
        type: integer
    required:
    - amountFen
    - id
    - name
    - points
    type: object
...
info:
  contact: {}
  description: REST API。HTTP 多为 200，业务结果看 JSON 的 code（0 成功）；鉴权接口需 Cookie session。
  title: AI Comic Generator API
  version: "1.0"
paths:
...
  /pay/admin/order/page:
    post:
      consumes:
      - application/json
      parameters:
      - description: 筛选与分页
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.AdminPayOrderPageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.PageResult'
              type: object
      security:
      - SessionCookie: []
      summary: 全站支付订单分页
      tags:
      - 支付管理
  /pay/admin/package/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: 方案
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.AddPayPackagePlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
            - properties:
                data:
                  type: integer
              type: object
      security:
      - SessionCookie: []
      summary: 新增充值方案
      tags:
      - 支付管理
  /pay/admin/package/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: id
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.DeletePayPackagePlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - SessionCookie: []
      summary: 删除充值方案
      tags:
      - 支付管理
  /pay/admin/package/page:
    post:
      consumes:
      - application/json
      parameters:
      - description: 分页
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.PayPackagePlanPageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.PageResult'
              type: object
      security:
      - SessionCookie: []
      summary: 管理端充值方案分页
      tags:
      - 支付管理
  /pay/admin/package/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: 方案
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.UpdatePayPackagePlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - SessionCookie: []
      summary: 更新充值方案
      tags:
      - 支付管理
  /pay/admin/receipt/page:
    post:
      consumes:
      - application/json
      parameters:
      - description: 筛选与分页
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.AdminPayOrderPageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.PageResult'
              type: object
      security:
      - SessionCookie: []
      summary: 收款记录分页（已支付）
      tags:
      - 支付管理
  /pay/mock-pay:
    post:
      consumes:
      - application/json
      parameters:
      - description: 订单号
        in: body
        name: body
        required: true
        schema:
          properties:
            orderNo:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - SessionCookie: []
      summary: 模拟支付成功
      tags:
      - 支付
...
  /pay/order:
    get:
      parameters:
      - description: 订单号
        in: query
        name: orderNo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
      security:
      - SessionCookie: []
      summary: 查询充值订单
      tags:
      - 支付
    post:
      consumes:
      - application/json
      parameters:
      - description: 套餐 code 等
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.CreatePayOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
      security:
      - SessionCookie: []
      summary: 创建充值订单
      tags:
      - 支付
  /pay/order/page:
    post:
      consumes:
      - application/json
      parameters:
      - description: 分页
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.PayOrderPageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.PageResult'
              type: object
      security:
      - SessionCookie: []
      summary: 我的充值订单分页
      tags:
      - 支付
  /pay/packages:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
      security:
      - SessionCookie: []
      summary: 充值套餐列表
      tags:
      - 支付
  /stat/dashboard:
    get:
      parameters:
      - description: 7d|30d|90d，默认 30d
        in: query
        name: range
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
      security:
      - SessionCookie: []
      summary: 管理端统计看板
      tags:
      - 统计
...
schemes:
- http
- https
securityDefinitions:
  SessionCookie:
    in: cookie
    name: session
    type: apiKey
swagger: "2.0"